refactor memory content models
